Move parsers to a separate file.
